django-app/requirements.txt
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11
django-cors-headers==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.8.0
gunicorn==19.7.1
icalendar==4.0.3
idna==2.7
itypes==1.1.0
Jinja2==2.10
lxml==4.2.5
MarkupSafe==1.1.0
openapi-codec==1.3.2
Pillow==5.3.0
psycopg2==2.7.4
PyPDF2==1.26.0
python-dateutil==2.7.5
pytz==2018.7
requests==2.20.1
simplejson==3.16.0
six==1.11.0
tika==1.19
uritemplate==3.0.0
urllib3==1.24.1
user-agent==0.1.9
